langchain/requirements.txt
langchain>=0.3.0
langchain-ollama>=0.3.0
python-dotenv>=1.0.0
